+++ test-utils/integration-test-utils-jms/src/main/java/uk/gov/justice/services/integrationtest/utils/jms/JmsMessageClientFactory.java
@@ -43,4 +43,10 @@ class JmsMessageClientFactory {
                 toJsonPathMessageConverter,
                 toJsonObjectMessageConverter);
     }
+
+    DefaultJmsAsyncMessageConsumerAsyncClient createJmsMessageConsumerAsyncClient() {
+        return new DefaultJmsAsyncMessageConsumerAsyncClient(jmsMessageConsumerPool,
+                jmsMessageReader,
+                toJsonEnvelopeMessageConverter);
+    }
 }