+++ test-utils/integration-test-utils-jms/src/main/java/uk/gov/justice/services/integrationtest/utils/jms/JmsMessageConsumerAsyncClient.java
@@ -1,0 +1,9 @@
+package uk.gov.justice.services.integrationtest.utils.jms;
+
+import uk.gov.justice.services.messaging.JsonEnvelope;
+
+import java.util.function.Consumer;
+
+public interface JmsMessageConsumerAsyncClient {
+    void registerOnMessageListener(final Consumer<JsonEnvelope> onMessageCallBack);
+}