+++ test-utils/integration-test-utils-jms/src/main/java/uk/gov/justice/services/integrationtest/utils/jms/JmsResourcesContext.java
@@ -49,6 +49,10 @@ public class JmsResourcesContext {
         jmsMessageConsumerPool.clearMessages();
     }
 
+    void closeAsyncConsumers() {
+        jmsMessageConsumerPool.closeAsyncConsumers();
+    }
+
     void closeConsumersAndProducers() {
         jmsMessageConsumerPool.closeConsumers();
         jmsMessageProducerFactory.closeProducers();